requirements.txt
APScheduler==3.6.3
blinker==1.9.0
cachetools==4.2.2
certifi==2025.1.31
chardet==5.2.0
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.57.0
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mysql-connector-python==9.0.0
numpy==2.2.4
packaging==24.2
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==13.15
pytz==2025.2
reportlab==4.3.1
six==1.17.0
tornado==6.1
tzdata==2025.2
tzlocal==5.3.1
Werkzeug==3.1.3
